Moved save() and associated methods from StlMesh to Mesh superclass. This allows both StlMesh and Mesh objects to save themselves, thereby allowing the creation of new stl files.
